Extract Oppiaine component
