+++ web/app/uusioppija/Oppiaine.jsx
@@ -1,0 +1,56 @@
+import React from 'baret'
+import Bacon from 'baconjs'
+import Atom from 'bacon.atom'
+import * as L from 'partial.lenses'
+import {UusiPerusopetuksenOppiaineDropdown} from '../perusopetus/UusiPerusopetuksenOppiaineDropdown'
+import {
+  accumulateModelState,
+  modelData,
+  modelLookup,
+  modelSet,
+  modelSetValues,
+  modelValid,
+  validateModel
+} from '../editor/EditorModel'
+import editorMapping from '../oppija/editors'
+import {Editor} from '../editor/Editor'
+import {PropertyEditor} from '../editor/PropertyEditor'
+import Text from '../i18n/Text'
+import Peruste from './Peruste'
+
+export default ({suoritusPrototypeP, oppiaineenSuoritusAtom, perusteAtom, oppilaitos, suorituskieli}) => { // suoritusPrototypeP = prototyyppi oppiaineen oppimäärän suoritukselle
+  return (<span>
+    {
+      Bacon.combineWith(suoritusPrototypeP, oppilaitos, suorituskieli, (oppiaineenSuoritus, toimipiste, kieli) => {
+        if (!oppiaineenSuoritus) return null
+
+        oppiaineenSuoritus = modelSetValues(Editor.setupContext(oppiaineenSuoritus, {edit:true, editorMapping}), {
+          toimipiste: { data: toimipiste },
+          suorituskieli: { data: kieli}
+        })
+
+        let oppiainePrototypeAtom = Atom(undefined) // Valittu oppiaine/koulutusmoduuliprototyyppi
+
+        let suoritusModelP = oppiainePrototypeAtom.flatMapLatest(oppiainePrototype => {
+          return oppiainePrototype && accumulateModelState(modelSet(oppiaineenSuoritus, oppiainePrototype, 'koulutusmoduuli'))
+        }).toProperty()
+
+        let suoritusDataP = suoritusModelP.map(model => model && modelValid(validateModel(model)) ? modelData(model) : null)
+
+        let suoritusP = Bacon.combineWith(suoritusDataP, perusteAtom, (suoritus, diaarinumero) => {
+          if (suoritus && diaarinumero) return L.set(L.compose('koulutusmoduuli', 'perusteenDiaarinumero'), diaarinumero, suoritus)
+        })
+
+        suoritusP.onValue(suoritus => oppiaineenSuoritusAtom.set(suoritus))
+
+        return (<span>
+          <Peruste suoritusTyyppiP={Bacon.constant(modelData(oppiaineenSuoritus, 'tyyppi'))} perusteAtom={perusteAtom} />
+          <label className="oppiaine"><Text name="Oppiaine"/>{' '}<UusiPerusopetuksenOppiaineDropdown oppiaineenSuoritus={oppiaineenSuoritus} selected={oppiainePrototypeAtom} resultCallback={s => oppiainePrototypeAtom.set(s)} pakollinen={true} enableFilter={false}/></label>
+          { suoritusModelP.map(model =>
+            model && <label><PropertyEditor model={modelLookup(model, 'koulutusmoduuli')} propertyName="kieli"/></label> )
+          }
+        </span>)
+      })
+    }
+  </span>)
+}
